+++ docs/en/changes/changes.md
@@ -20,6 +20,7 @@
 * Add Ztunnel component in the topology.
 * [Break Change] Change `compomentId` to `componentIds` in the K8SServiceRelation Scope. 
 * Adapt the mesh metrics if detect the ambient mesh in the eBPF access log receiver.
+* Add JSON format support for the `/debugging/config/dump` status API.
 
 #### UI